Fixes Bug when generating ecr App
Currently when generating the app with option -t ecr the controller
still defaults to the slang template enigine.

This fixes this bug by dynamically passing the template engine extension
to the controller render method.

With this change users should be able to generate and run amber apps
with slang or ecr template engine.